refactor: Make analyze pubspecs methods reusable and only hold one pubspec in memory at a time
